move authorizeUser to middleware
